requirements.txt
python-dotenv
jupyterlab
ipywidgets
urllib3
torch
transformers
accelerate
sentencepiece
bitsandbytes
tqdm
openai
gradio
langchain
tiktoken
faiss-cpu
langchain-openai
langchain_experimental
langchain_chroma
langchain[docarray]
datasets
matplotlib
google.generativeai
anthropic
scikit-learn
unstructured
chromadb
plotly
jupyter-dash
groq